fix(dst): preserve host NumCPU range
Spec-first: canonical=docs/dst/design.md deterministic process identity and docs/dst/faults.md per-host identity; contract=every accepted positive NumCPU is observed exactly; mechanism=store per-host NumCPU as the public target-width int; collapse-check=faithful.

Diagnosis: fault=positive HostConfig.NumCPU values above MaxInt32 are observed as defaults or unrelated counts; proximate=dstSetHostIdent narrowed int to int32; root=accepted identity values were not represented losslessly; fixing at=runtime host identity representation; wider scope justified by=n/a — narrowest sufficient.

Invariant: kind=entailed; property=positive accepted NumCPU round-trips exactly; from=identity determinism and consistency; violation=MaxInt32+1 is accepted but runtime.NumCPU reports the run default.

Structural-check: invariant=accepted NumCPU is observed exactly; illegal-state=public int differs after publication; representable-because=redundant/derivable narrowing to int32; decision=collapse to int in dstHostIdentity.

Triage at 57.1: 1 match; folding dst-host-numcpu-wraps-int32, redeferring none, closing none.

Spec: docs/dst/design.md and docs/dst/faults.md; Plan: docs/plans/dst-audit-issue-burn-down.md; divergences: none.

+++ src/runtime/dst.go
@@ -1411,7 +1411,7 @@ var (
 // processes share a hostname/NumCPU while different hosts can differ.
 type dstHostIdentity struct {
 	hostname string
-	numcpu   int32
+	numcpu   int
 	set      bool
 }
 
@@ -1544,7 +1544,7 @@ func dstSetHostIdent(host uint32, hostname string, numcpu int) {
 		if old != nil {
 			copy(ent, old.ent)
 		}
-		ent[host] = dstHostIdentity{hostname: hostname, numcpu: int32(numcpu), set: true}
+		ent[host] = dstHostIdentity{hostname: hostname, numcpu: numcpu, set: true}
 		if dstHostIdent.CompareAndSwap(old, &dstHostIdentTable{ent: ent}) {
 			return
 		}
